requirements.txt
discord.py==2.4.0
Pillow==10.3.0
